create simbricks.orchestration python package (#14)
I basically did the following:
`mkdir experiments/simbricks/orchestration`
`mv experiments/simbricks/* experiments/simbricks/orchestration/`

Additionally, imports and paths had to be adjusted.
