+++ experiments/pyexps/ae/f7_scale.py
@@ -42,10 +42,10 @@
 # $: python3 run.py pyexps/ae/f7_scale.py --filter host-gt-ib-sw-* --verbose
 ########################################################################
 
-import simbricks.experiments as exp
-import simbricks.nodeconfig as node
-import simbricks.simulators as sim
-from simbricks.simulator_utils import create_basic_hosts
+import simbricks.orchestration.experiments as exp
+import simbricks.orchestration.nodeconfig as node
+import simbricks.orchestration.simulators as sim
+from simbricks.orchestration.simulator_utils import create_basic_hosts
 
 host_types = ['gt']
 nic_types = ['ib']
