+++ experiments/pyexps/dist_memcache.py
@@ -23,11 +23,11 @@
 import math
 import random
 
-import simbricks.experiments as exp
-import simbricks.nodeconfig as node
-import simbricks.simulators as sim
-from simbricks import proxy
-from simbricks.simulator_utils import create_multinic_hosts
+import simbricks.orchestration.experiments as exp
+import simbricks.orchestration.nodeconfig as node
+import simbricks.orchestration.simulators as sim
+from simbricks.orchestration import proxy
+from simbricks.orchestration.simulator_utils import create_multinic_hosts
 
 host_types = ['qemu', 'gem5', 'qt']
 n_nets = [1, 2, 3, 4, 8, 16, 32]
