+++ experiments/simbricks/orchestration/experiment/experiment_output.py
@@ -20,7 +20,49 @@
 # TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
 # SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-from simbricks.runtime.common import Run, Runtime
-from simbricks.runtime.distributed import DistributedSimpleRuntime, auto_dist
-from simbricks.runtime.local import LocalParallelRuntime, LocalSimpleRuntime
-from simbricks.runtime.slurm import SlurmRuntime
+import json
+import time
+
+from simbricks.orchestration.experiments import Experiment
+
+
+class ExpOutput(object):
+    """Manages an experiment's output."""
+
+    def __init__(self, exp: Experiment):
+        self.exp_name = exp.name
+        self.metadata = exp.metadata
+        self.start_time = None
+        self.end_time = None
+        self.sims = {}
+        self.success = True
+        self.interrupted = False
+
+    def set_start(self):
+        self.start_time = time.time()
+
+    def set_end(self):
+        self.end_time = time.time()
+
+    def set_failed(self):
+        self.success = False
+
+    def set_interrupted(self):
+        self.success = False
+        self.interrupted = True
+
+    def add_sim(self, sim, comp):
+        obj = {
+            'class': sim.__class__.__name__,
+            'cmd': comp.cmd_parts,
+            'stdout': comp.stdout,
+            'stderr': comp.stderr,
+        }
+        self.sims[sim.full_name()] = obj
+
+    def dumps(self):
+        return json.dumps(self.__dict__)
+
+    def loads(self, json_s):
+        for k, v in json.loads(json_s).items():
+            self.__dict__[k] = v
