+++ experiments/simbricks/orchestration/simulator_utils.py
@@ -23,9 +23,11 @@
 
 import typing as tp
 
-from simbricks.experiments import Experiment
-from simbricks.nodeconfig import AppConfig, NodeConfig
-from simbricks.simulators import HostSim, I40eMultiNIC, NetSim, NICSim
+from simbricks.orchestration.experiments import Experiment
+from simbricks.orchestration.nodeconfig import AppConfig, NodeConfig
+from simbricks.orchestration.simulators import (
+    HostSim, I40eMultiNIC, NetSim, NICSim
+)
 
 
 def create_basic_hosts(
